requirements.txt
customtkinter
pillow
tkcalendar
python-dateutil
pyserial
